+++ server/src/test/kotlin/io/spine/server/route/ProjectionClassRoutingSpec.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022, TeamDev. All rights reserved.
+ * Copyright 2023, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,24 +23,17 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-syntax = "proto3";
 
-package spine.server.given.sorting;
+package io.spine.server.route
 
-import "spine/options.proto";
+import org.junit.jupiter.api.DisplayName
+import org.junit.jupiter.api.Test
 
-option (type_url_prefix) = "type.spine.io";
-option java_package="io.spine.server.given.sorting.command";
-option java_outer_classname = "SortingCommandsProto";
-option java_multiple_files = true;
+@DisplayName("`ProjectionClass` routing should")
+class ProjectionClassRoutingSpec {
 
-import "spine/server/given/sorting/figure.proto";
-
-// A request to generate geometric figures specified in the [figure] field of the command.
-//
-// The generation is repeated [count] times.
-//
-message GenerateFigures {
-    repeated Figure figure = 1 [(required) = true, (distinct) = true];
-    int32 count = 2 [(min).value = "1"];
+    @Test
+    fun `provide routing via annotated static method`() {
+        // TODO implement.
+    }
 }
