+++ server/src/testFixtures/proto/spine/server/given/context/sorting/sorter.proto
@@ -25,20 +25,38 @@
  */
 syntax = "proto3";
 
-package spine.server.given.sorting;
+package spine.server.given.context.sorting;
 
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.spine.io";
-option java_package="io.spine.server.given.sorting";
-option java_outer_classname = "FigureProto";
+option java_package="io.spine.server.given.context.sorting";
+option java_outer_classname = "SorterProto";
 option java_multiple_files = true;
 
-// A geometry figure.
-message Figure {
-    oneof kind {
-        bool circle = 1;
-        bool triangle = 2;
-        bool square = 3;
+import "google/protobuf/empty.proto";
+
+import "spine/server/given/context/sorting/figure.proto";
+
+// Accumulates figures by buckets.
+message Sorter {
+    option (entity).kind = VIEW;
+
+    // The sorter is a singleton.
+    google.protobuf.Empty id = 1;
+
+    // Use `repeated` wince we cannot have have a map key of type [Figure].
+    repeated Bucket bucket = 2;
+
+    // A container for [Figure]s.
+    message Bucket {
+        repeated Figure figure = 3;
     }
+}
+
+// Counts how many figures were generated.
+message FigureStats {
+    option (entity).kind = VIEW;
+    Figure figure = 1;
+    int32 count = 2 [(min).value = "0"];
 }
